Bugfix for high frequency stimulation:
For stimulation frequencies over 60Hz or low end hardware, the stimulator responce for a pulse might arrive after the next pulse is send.
This is probably a FTDI driver issue. However, the error could lead to a race condition and block for ever.

The output for timing errors was also improved.
A new configuration variable was introduced to disable timing error messages.